ajout UserPolicy, modif code PostPolicy, ajout messages de validation en français pour le mot de passe + amélioration messages des requests, bug StoreUserRequest corrigé

+++ app/Http/Controllers/API/UserController.php
@@ -84,6 +84,8 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        // policy pour vérifier que l'utilisateur peut modifier le compte
+        $this->authorize('update', $user);
 
         // On modifie les informations de l'utilisateur
         $user->update([
@@ -132,6 +134,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        // policy pour vérifier que l'utilisateur peut supprimer le compte
+        $this->authorize('delete', $user);
+
         // on supprimer l'utilisateur en base de données
         $user->delete();
